backend/requirements.txt
# Core Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
neomodel>=5.3.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Dev & Test
ruff>=0.6.0
pytest>=8.3.0
httpx>=0.27.0

# Agents (Sprint 2) - install separately:
# pip install langchain langchain-google-genai langgraph
